Having the necessary changes made to be able to get things moving forward
