+++ Startup.cs
@@ -66,6 +66,11 @@ namespace BotDontLie
             // Adding the necessary helper classes as well.
             services.AddSingleton<ITeamHelpers, TeamHelpers>();
 
+            services.AddSingleton<IDataHelpers, DataHelpers>((provider) => new DataHelpers(
+                provider.GetRequiredService<IBallDontLieService>(),
+                provider.GetRequiredService<TelemetryClient>(),
+                provider.GetRequiredService<ITeamHelpers>()));
+
             // Adding the games provider.
             services.AddSingleton<IGamesProvider, GamesProvider>((provider) => new GamesProvider(
                 this.Configuration["StorageConnectionString"],
@@ -88,9 +93,8 @@ namespace BotDontLie
             // Create the bot as a transient. In this case the ASP Controller is expecting an IBot.
             services.AddTransient<IBot, NbaBot>((provider) => new NbaBot(
                 provider.GetRequiredService<TelemetryClient>(),
-                provider.GetRequiredService<IBallDontLieService>(),
                 this.Configuration["AppBaseUri"],
-                provider.GetRequiredService<ITeamHelpers>()));
+                provider.GetRequiredService<IDataHelpers>()));
 
             // Adding the HttpClient.
             services.AddHttpClient("BallDontLieAPI", c =>
